requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# plugin\archive.py: 3
archiveis == 0.0.6

# requirements.py: 2
# variables.py: 1
cryptography == 2.3.1

# Main.py: 1,2
# plugin\aloh.py: 4,6
# plugin\archive.py: 5
# plugin\crypto.py: 3
# plugin\etc.py: 1,3
# plugin\exec.py: 5,7
# plugin\gravatar.py: 3
# plugin\mod.py: 4
# plugin\osu.py: 4,5
# plugin\post.py: 3,5
# plugin\tag.py: 7
# plugin\whisperandlog.py: 4,6
discord.py == 0.16.12

# plugin\osu.py: 1
osuapi == 0.0.33

# plugin\aloh.py: 2
# plugin\mod.py: 1
# plugin\post.py: 2
# plugin\tag.py: 2
# plugin\whisperandlog.py: 3
psycopg2_binary == 2.7.5

# plugin\archive.py: 6,7,8,9,10
selenium == 3.6.0
